Move WinterAreasQuery component to query folder
