+++ src/components/query/WinterAreasQuery.tsx
@@ -1,5 +1,5 @@
 import { Query } from 'react-apollo';
-import { WinterAreasQuery as WinterAreasQueryTypes } from '../../../utils/__generated__/WinterAreasQuery';
+import { WinterAreasQuery as WinterAreasQueryTypes } from '../../utils/__generated__/WinterAreasQuery';
 
 class WinterAreasQuery extends Query<WinterAreasQueryTypes> {}
